refactor: changed name of AppStyles to the styled component it contains
