+++ frontend/src/styles/ContentWindow.tsx
